Improve EOF error messaging during CLI login flows
Motivated by observed issues in Compute - users can set up working
Compute endpoints, change their setups slightly around access to STDIN,
and then see cryptic EOF errors when they eventually need to re-auth.

NB: as of the time of this commit, Compute handles this by raising an
error before running login flows if the current environment does not
appear to be an interactive terminal. The goal of this commit is to move
that behavior to a more centralized location.